frontend/requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.6
markdown2==2.4.13
MarkupSafe==2.1.5
packaging==24.1
pillow==10.3.0
pytesseract==0.3.10
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
pymupdf